Backend/Data/requirements.txt
# Dependencias para los scripts de carga de datos
datasets>=2.0.0
pandas>=1.5.0
Werkzeug>=2.0.0
kagglehub[pandas-datasets]>=0.2.0
